STK: Snapshot 10-07-24 12:22 from Sierra 5.21.5-699-g38edc8e6
