+++ packages/stk/stk_unit_test_utils/stk_unit_test_utils/ElemGraphMultipleSharedSidesUtils.hpp
@@ -56,6 +56,7 @@
 #include <stk_mesh/base/SkinMesh.hpp>
 #include <stk_mesh/baseImpl/elementGraph/ElemElemGraph.hpp>
 #include <stk_mesh/baseImpl/elementGraph/ElemElemGraphImpl.hpp>
+#include <stk_mesh/baseImpl/elementGraph/ProcessKilledElements.hpp>
 
 #include <stk_util/parallel/Parallel.hpp>
