+++ packages/stk/stk_unit_tests/stk_mesh/face_creation/element_graph/UnitTestElementDeath.cpp
@@ -10,6 +10,7 @@
 #include <stk_mesh/base/GetEntities.hpp>  // for get_selected_entities, etc
 #include <stk_mesh/base/MetaData.hpp>   // for MetaData
 #include <stk_mesh/baseImpl/elementGraph/ElemElemGraph.hpp>  // for process_killed_elements, etc
+#include <stk_mesh/baseImpl/elementGraph/ProcessKilledElements.hpp>
 #include <stk_topology/topology.hpp>    // for topology, etc
 #include <stk_unit_test_utils/ioUtils.hpp>  // for fill_mesh_using_stk_io, etc
 #include <stk_util/parallel/Parallel.hpp>  // for parallel_machine_size, etc
