+++ packages/stk/stk_unit_tests/stk_mesh/face_creation/integration/UnitTestSkinMeshElementDeath.cpp
@@ -5,6 +5,7 @@
 #include <stk_io/IossBridge.hpp>        // for put_io_part_attribute
 #include <stk_mesh/base/BulkData.hpp>   // for BulkData
 #include <stk_mesh/baseImpl/elementGraph/ElemElemGraph.hpp>  // for process_killed_elements, etc
+#include <stk_mesh/baseImpl/elementGraph/ProcessKilledElements.hpp>
 #include <stk_mesh/base/GetEntities.hpp>  // for get_selected_entities
 #include <stk_mesh/base/MetaData.hpp>   // for MetaData
 #include <stk_topology/topology.hpp>    // for topology, etc
